+++ src/docs/guide/modules/serverless/example_gcf.kt
@@ -1,11 +1,16 @@
 package guide.modules.serverless
 
-import org.http4k.core.*
+import org.http4k.core.Filter
+import org.http4k.core.HttpHandler
+import org.http4k.core.Method
+import org.http4k.core.Request
+import org.http4k.core.Response
 import org.http4k.core.Status.Companion.INTERNAL_SERVER_ERROR
 import org.http4k.core.Status.Companion.OK
+import org.http4k.core.then
 import org.http4k.routing.bind
 import org.http4k.routing.routes
-import org.http4k.serverless.gcp.Http4kGCFAdapter
+import org.http4k.serverless.gcf.GoogleCloudFunction
 
 val routes = routes(
     "/success" bind Method.GET to { request: Request -> Response(OK).body("works!") },
@@ -21,4 +26,4 @@ val timer = Filter { next: HttpHandler ->
     }
 }
 
-class FunctionsExampleEntryClass : Http4kGCFAdapter(timer.then(routes))
+class FunctionsExampleEntryClass : GoogleCloudFunction(timer.then(routes))