+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This list is not currently intended to be all-encompassing - it will document ma
 changes with their rationale when appropriate:
 
 ### v3.249.0 (uncut)
+- [all] Upgrade some dependency versions.
 - [http4k-server-netty] Add support for response streaming.  H/T @carbotaniuman for the PR.
+- [http4k-serverless-gcf] New serverless module! H/T @ssijak for the PR.
 
 ### v3.248.0
 - [http4k-server-ratpack] New backend module!